UI: Refactor MyUser lens file structure
Change-type: patch
